requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
choicelib==0.1.5
colorama==0.4.6
dotenv==0.9.9
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
vbml==1.1.post1
vk-api==11.9.9
vkbottle==4.5.1
vkbottle-types==5.199.93.4.post1
yarl==1.20.0
